Rename product_hpd_matrices_and_siegel_disks into product_hpd_and_siegel_disks in file names
