+++ geomstats/geometry/product_hpd_and_siegel_disks.py
